Changes: - Refactorized Meter class, removing unnecessary members, setting default values ​​and checking values ​​to set - Implemented but disabled/commented the DropDigits feature - Implemented but disabled/commented the ArtificialTimeSync feature - Increased app version to 1.39.05

+++ Xml/MeterTypes.cs
@@ -70,8 +70,8 @@ namespace Xml
         }
 
         public List<Meter> FindByEncoderTypeAndLiveDigits (
-            int encoderType,
-            int liveDigits )
+            byte encoderType,
+            int  liveDigits )
         {
             // Filter by protocol and livedigits
             List<Meter> meters = Meters.FindAll ( x => (
